Moves resource classes into separate files
